feat(shipping-extensions): file rename
